Move WalletResolver into integrations — it is plugin-facing glue
The resolver composes the accounts and custody domains and exists solely
to populate IntegrationContext.walletResolver for the deposit plugins
(wallet-deposit consumes it as-is; collateral adapts it, and substitutes
a credentials-registry implementation in contract mode). It was also the
only custody file importing from accounts/, breaking the otherwise
one-way accounts -> custody graph. Relocate next to the registry that
declares its context field; the function-type contract stays — it is the
narrowest port plugins need and is already mode-swapped.
