+++ src/integrations/deposits/collateral/index.ts
@@ -4,7 +4,7 @@ import {
   WalletResolver as CollateralWalletResolver,
 } from "@owneraio/finp2p-ethereum-collateral";
 import { FinP2PContract } from "@owneraio/finp2p-contracts";
-import { WalletResolver as CustodyWalletResolver } from "../../../services/custody";
+import { WalletResolver as CustodyWalletResolver } from "../../wallet-resolver";
 import { IntegrationContext } from "../../registry";
 import { pooledProvider, pooledSigner } from "../../signer-pool";
