+++ src/integrations/deposits/wallet-deposit/plugin.ts
@@ -12,7 +12,7 @@ import {
   successfulDepositOperation,
   failedDepositOperation,
 } from "@owneraio/finp2p-nodejs-skeleton-adapter";
-import { WalletResolver } from "../../../services/custody";
+import { WalletResolver } from "../../wallet-resolver";
 import { AssetStore } from "../../../services/accounts";
 import { paymentsSlotClaimedExternally } from "../payments-slot";
 import { IntegrationContext } from "../../registry";
