+++ src/integrations/registry.ts
@@ -3,7 +3,8 @@ import { PluginManager } from "@owneraio/finp2p-nodejs-skeleton-adapter";
 import { InboundTransferHook } from "@owneraio/finp2p-nodejs-skeleton-adapter/plugin";
 import { FinP2PClient } from "@owneraio/finp2p-client";
 import { FinP2PContract } from "@owneraio/finp2p-contracts";
-import { CustodyProvider, WalletResolver } from "../services/custody";
+import { CustodyProvider } from "../services/custody";
+import { WalletResolver } from "./wallet-resolver";
 import { AssetStore } from "../services/accounts";
 import { AccountModel } from "../config";
 import { registerFireblocks } from "./fireblocks";
