+++ src/services/custody/index.ts
@@ -2,5 +2,4 @@ export * from './custody-provider'
 export * from './custody-registry'
 export * from './gas-station'
 export * from './wallet-activation'
-export * from './wallet-resolver'
 export * from './health'
